WIP: details row check
