+++ packages/office-ui-fabric-react/src/components/Check/Check.base.tsx
@@ -1,10 +1,9 @@
 import * as React from 'react';
-import { BaseComponent, classNamesFunction, customizable } from '../../Utilities';
-import { ICheckProps, ICheckStyleProps, ICheckStyles } from './Check.types';
+import { BaseComponent, customizable } from '../../Utilities';
+import { ICheckProps } from './Check.types';
 import { Icon } from '../../Icon';
 import { getStyles } from './Check.styles';
-
-const getClassNames = classNamesFunction<ICheckStyleProps, ICheckStyles>();
+import { getClassNames } from './Check.classNames';
 
 @customizable('Check', ['theme', 'styles'])
 export class CheckBase extends BaseComponent<ICheckProps, {}> {
