+++ packages/office-ui-fabric-react/src/components/Check/Check.classNames.ts
@@ -1,0 +1,4 @@
+import { classNamesFunction } from '../../Utilities';
+import { ICheckStyleProps, ICheckStyles } from './Check.types';
+
+export const getClassNames = classNamesFunction<ICheckStyleProps, ICheckStyles>();
