+++ packages/office-ui-fabric-react/src/components/Check/Check.styles.ts
@@ -54,7 +54,7 @@ export const getStyles = (props: ICheckStyleProps): ICheckStyles => {
            * TODO: Come back to this once .checkHost has been
            * converted to mergeStyles
            */
-          '.checkHost:hover &, .checkHost:focus &, &:hover, &:focus': {
+          '$checkHost:hover &, $checkHost:focus &, &:hover, &:focus': {
             opacity: 1
           }
         }
@@ -131,6 +131,8 @@ export const getStyles = (props: ICheckStyleProps): ICheckStyles => {
           }
         }
       }
-    ]
+    ],
+
+    checkHost: [{}]
   };
 };
