+++ packages/office-ui-fabric-react/src/components/Check/Check.types.ts
@@ -72,4 +72,9 @@ export interface ICheckStyles {
    * The 'circle' icon styles.
    */
   circle: IStyle;
+
+  /**
+   * Check host style
+   */
+  checkHost: IStyle;
 }
